task 3/pm/requirements.txt
psycopg2-binary==2.9.9
Pillow==10.1.0
